RelativeIndex rank field type changed to double

+++ include/search_server.h
@@ -23,7 +23,7 @@ struct DocRelevance {
 
 struct RelativeIndex {
     size_t doc_id;
-    float rank;
+    double rank;
 
     bool operator==(const RelativeIndex &other) const {
         return (doc_id == other.doc_id && rank == other.rank);
